==> Emma_Davies/1. combining_data.ipynb <==
import pandas as pd
import os
print(os.getcwd())

#read BSc data file and name it 'df1'

df1 = pd.read_csv("cross_solvent_solubility_dataset_BSc_2024.csv")

print(df1)

#define LogS

logS = df1["LogS"]

#replace logS with logS_M

df1.rename(columns={"LogS": "logS_M"}, inplace=True)

# Remove the 'ID' column from the DataFrame
df1.drop(columns=['ID'], inplace=True)

#read basf file and name it df2

df2 = pd.read_csv("basf_inchi_logS.csv")

print(df2)

#define solute inchikey column

solute_inchikey = df2["solute_inchikey"]

#remove inchikey from the solute_inchikey column

solute_inchikey_edited = solute_inchikey.str.replace("InChIKey=", "", regex=False)

print(solute_inchikey_edited)

#replace the column in the data2

df2["solute_inchikey"] = solute_inchikey_edited

# Remove the 'MP_C' column from the data2
df2.drop(columns=['MP_C'], inplace=True)

print(df2)

#combine data files and remove duplicates

#check columns in df1 and df2
print("df1 columns:", df1.columns)
print("df2 columns:", df2.columns)

#check format of columns
df1.columns = df1.columns.str.strip()
df2.columns = df2.columns.str.strip()

df1['solute_name'] = df1['solute_name'].astype(str)
df2['solute_name'] = df2['solute_name'].astype(str)


# Combine the two data files
combined_df = pd.concat([df1, df2])

# Save the combined data to a new CSV file (if needed)
combined_df.to_csv('combined_data.csv', index=False)

combined_data = pd.read_csv('combined_data.csv')
print(combined_data)

#check column names
print(combined_data.columns)

# Identify duplicate rows except for 'LogS_M'
columns_to_check = [col for col in combined_data.columns if col not in ['logS_M', 'MW', 'S_M', 'solute_inchikey']]
duplicates = combined_data[combined_data.duplicated(subset=columns_to_check, keep=False)].copy()

# Save duplicate rows
duplicates.to_csv('combined_data_duplicates.csv', index=False)

print(duplicates.shape)  # Check how many duplicates were identified
print(combined_data.shape)  # Check original data size

# Remove original duplicate rows from the dataframe
combined_data_no_duplicates = combined_data.drop(duplicates.index)
combined_data_no_duplicates.to_csv('combined_data_no_duplicates.csv', index=False)

# Ensure 'LogS' is numeric before averaging
duplicates.loc[:, 'logS_M'] = pd.to_numeric(duplicates['logS_M'], errors='coerce').fillna(0)
duplicates.loc['MW'] = pd.to_numeric(duplicates['MW'], errors='coerce')
duplicates.loc['S_M'] = pd.to_numeric(duplicates['S_M'], errors='coerce')

# Compute the average 'LogS' for duplicate groups
averaged_duplicates = duplicates.groupby(columns_to_check + ['solute_inchikey'], as_index=False).agg({'logS_M': 'mean', 'MW': 'mean', 'S_M':'mean'})

# Append the averaged rows
combined_data_no_duplicates = pd.concat([combined_data_no_duplicates, averaged_duplicates], ignore_index=True)

# Save the final dataframe
combined_data_no_duplicates.to_csv('combined_data_no_duplicates_and_ave_duplicates.csv', index=False)

#print combined_data_no_duplicates_and_ave_duplicates
df = pd.read_csv('combined_data_no_duplicates_and_ave_duplicates.csv')
print(df)
