==> Emma_Davies/3. data_coverage.ipynb <==
import pandas as pd

# Load the CSV file
df = pd.read_csv("filtered_data_updated_final2.csv")

# Count unique entries
unique_solutes = df['solute_inchikey'].nunique()
unique_solvents = df['solvent_name'].nunique()

print(f"Number of unique solutes: {unique_solutes}")
print(f"Number of unique solvents: {unique_solvents}")

# Calculate total number of possible combinations
combinations = 319*77
percentage = len(df['logS_M'])*100/combinations
print ('Number of possible combinations of solute and solvent: ', combinations)
print ('Percentage of data coverage: ', percentage, '%')
